Audit + CleanUp
git-svn-id: https://cchlib.googlecode.com/svn/trunk@8254 bc589049-105d-1923-a3ce-6fcb24364ed1

+++ cchlib-i18n-deprecated/src/main/java/com/googlecode/cchlib/i18n/I18nSimpleStatsResourceBundle.java
@@ -42,7 +42,7 @@ public class I18nSimpleStatsResourceBundle extends I18nSimpleResourceBundle
             map.put( key, 1 );
             }
         else {
-            map.put( key, count + 1 );
+            map.put( key, count.intValue() + 1 );
             }
 
         return super.getString( key );
